+++ src/frontend/web_application/src/scenes/NewDraft/presenter.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import NewDraftForm from '../../components/NewDraftForm';
+import DraftMessageActionsContainer from '../../components/DraftMessageActionsContainer';
 
 class NewDraft extends Component {
   static propTypes = {
@@ -9,8 +10,9 @@ class NewDraft extends Component {
     internalId: PropTypes.string,
     requestNewDraft: PropTypes.func.isRequired,
     editDraft: PropTypes.func.isRequired,
-    saveDraft: PropTypes.func.isRequired,
     sendDraft: PropTypes.func.isRequired,
+    onSaveDraft: PropTypes.func.isRequired,
+    onDeleteMessage: PropTypes.func.isRequired,
   };
 
   static defaultProps = {
@@ -32,16 +34,34 @@ class NewDraft extends Component {
     return action({ internalId, draft, message });
   };
 
+  handleDelete = () => {
+    const { message, onDeleteMessage } = this.props;
+    onDeleteMessage({ message });
+  };
+
+  renderDraftMessageActionsContainer = () => {
+    const { message, internalId } = this.props;
+
+    return (
+      <DraftMessageActionsContainer
+        message={message}
+        internalId={internalId}
+        onDelete={this.handleDelete}
+      />
+    );
+  }
+
   render() {
-    const { draft, internalId, editDraft, saveDraft, sendDraft } = this.props;
+    const { draft, internalId, editDraft, onSaveDraft, sendDraft } = this.props;
 
     return (
       <NewDraftForm
         internalId={internalId}
         draft={draft}
         onChange={this.makeHandle(editDraft)}
-        onSave={this.makeHandle(saveDraft)}
+        onSave={this.makeHandle(onSaveDraft)}
         onSend={this.makeHandle(sendDraft)}
+        renderDraftMessageActionsContainer={this.renderDraftMessageActionsContainer}
       />
     );
   }